Moved ResultSet class to internal util package
